move functionality and only keep levels functionality
