+++ ehr2meds/meds_stages/apply_adaptive_code_mapping.py
@@ -4,10 +4,21 @@ from __future__ import annotations
 
 import polars as pl
 from collections.abc import Callable
-from ehr2meds.meds_stages.adaptive_code_mapping import apply_mapping, prepare_mapping
+from ehr2meds.adaptive_code_mapping import CODE_COLUMN, MAPPED_CODE_COLUMN, prepare_mapping
 from MEDS_transforms.stages import Stage
 from omegaconf import DictConfig
 from pathlib import Path
+
+
+def apply_mapping(data: pl.LazyFrame, mapping: pl.DataFrame) -> pl.LazyFrame:
+    """Rewrite codes through a frozen mapping while preserving row order and schema."""
+    temporary = "_adaptive_mapped_code"
+    lookup = mapping.select(CODE_COLUMN, pl.col(MAPPED_CODE_COLUMN).alias(temporary)).lazy()
+    return (
+        data.join(lookup, on=CODE_COLUMN, how="left", maintain_order="left")
+        .with_columns(pl.coalesce(temporary, CODE_COLUMN).alias(CODE_COLUMN))
+        .drop(temporary)
+    )
 
 
 @Stage.register(
